move related collections out of tree generation
